Correcetly simulate passed time before first clock edge in verilator
The FFI code used a default of 50 nanoseconds as time passed before the first clock edge.
This resulted in an unobservable reset pulse when used with `resetGenN d1` because  verilator's simulation time is off by a factor 1000.

Now derive the time before the clock edge as a single period of vexrisc clock domain.